requirements.txt
docker
django
djangorestframework
djangorestframework-jwt
gunicorn
kubernetes
django-cors-headers